feat(landing): update documentation and components for API keys integration
