+++ apps/landing/content/docs/1.getting-started/1.index.md
@@ -5,8 +5,6 @@ navigation:
   icon: lucide:book-open
 ---
 
-:docs-hero
-
 Shotly is a modern platform for organizing your bookmarks. It helps you save, categorize, and quickly find your favorite links — whether for work, study, or inspiration.
 
 With Shotly, you can create collections, use tags, and benefit from instant search. Import and export your bookmarks, manage them privately, and access your collection from any device. The intuitive interface and fast performance ensure a smooth experience for both individuals and teams.
