+++ apps/landing/content/docs/4.advanced/api-keys.md
@@ -1,0 +1,27 @@
+---
+title: API Keys
+description: Generate API keys to access Shotly API.
+navigation:
+  icon: lucide:key-round
+---
+
+:docs-image{light="/docs/advanced/api-keys/cover-light.png" dark="/docs/advanced/api-keys/cover-dark.png"}
+
+## Overview
+
+API keys provide a secure and controlled way to access Shotly's functionality for external applications and automation. With them, you can integrate third-party services, scripts, or tools without exposing your credentials. This is especially useful for automating routine tasks, creating custom integrations, or extending the platform's capabilities through external applications.
+
+## Create API key
+
+1. Go to the Shotly settings and select the **API keys** section.
+2. In the "Create API key" block, enter a unique name for the new key — this will help distinguish keys from each other.
+3. Specify the key expiration date (e.g., 1 day, 1 month, 12 months, or no expiration).
+4. Click the **Create API key** button.
+5. After creation, copy the generated key and save it in a secure place. For security reasons, the key will be displayed only once.
+
+## Usage
+
+API keys can be used for:
+- Integration with external services and automation (e.g., automatically adding links through scripts);
+- Getting statistics and managing collections through the API;
+- Creating your own tools and extensions for Shotly.
