--- apps/landing/content/docs/4.advanced/export.md
@@ -1,10 +1,0 @@
----
-title: API Keys
-description: Export your Shotly data for backup or migration.
-navigation:
-  icon: lucide:key-round
----
-
-Exporting Shotly data is designed for creating backups and transferring information to other systems. This is an essential tool for ensuring the safety and integrity of user data.
-
-The export capability allows integration with external services and automates information exchange processes between different platforms.
