+++ pnpm-workspace.yaml
@@ -19,3 +19,4 @@ ignoredBuiltDependencies:
 
 onlyBuiltDependencies:
   - better-sqlite3
+  - sharp
